Menu can be displayed and use input handled
The displayMenu subroutine is used to display the menu options to the user and the number related to each option. The getMenuChoice subroutine is used to get the number from the user. This input is then handled by checking if it is a number for an option that appears in the menu. If it is, that option will be carried out (code for options 1 and 2
not written yet). If not, an error message will be displayed using the displayError subroutine which formats and displays an error to the user.